backend/requirements.txt
# Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket
websockets==12.0
wsproto==1.2.0

# SSH
asyncssh==2.14.2

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Authentication & Security
bcrypt==4.0.1
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6

# Utility
python-dotenv==1.0.0
cryptography==41.0.7